Avoid failures if latest nightly DPCPP tag didn't provide binaries  (#426)
On certain conditions a nightly tag for DPCPP may exist without
providing the binary packages for linux. This artifact can be found in
the API description for the release tag, under the `assets` section.

This PR filters release tags and avoids the ones without the binary
package `sycl_linux.tar.gz`

It also reverts the workaround to avoid this failure during the day it
was discovered.

---------

Co-authored-by: Alejandro Acosta <alejandro.acosta@codeplay.com>